# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Course 后端 API 服务",
        "title": "AI Course API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/grading/assignment/{assignment_id}/progress": {
            "get": {
                "description": "获取作业的批改进度统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业批改"
                ],
                "summary": "获取批改进度",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "assignment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/grading/assignment/{assignment_id}/publish": {
            "post": {
                "description": "教师发布作业成绩，学生可查看批改结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业批改"
                ],
                "summary": "发布成绩",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "assignment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发布成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/grading/assignment/{assignment_id}/submissions": {
            "get": {
                "description": "教师获取指定作业的待批改提交列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业批改"
                ],
                "summary": "获取待批改提交列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "assignment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "submitted",
                            "graded"
                        ],
                        "type": "string",
                        "description": "提交状态过滤",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/grading/batch": {
            "post": {
                "description": "教师对多个提交进行批量快速批改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业批改"
                ],
                "summary": "批量快速批改",
                "parameters": [
                    {
                        "description": "批量批改信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BatchGradeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批改成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/grading/submission/{submission_id}": {
            "get": {
                "description": "教师获取特定提交的批改详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业批改"
                ],
                "summary": "获取批改详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "提交ID",
                        "name": "submission_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "提交不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "教师批改学生提交的作业",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业批改"
                ],
                "summary": "批改提交",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "提交ID",
                        "name": "submission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "批改信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GradeSubmissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "批改成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "提交不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.BatchGradeItem": {
            "type": "object",
            "required": [
                "answers",
                "submission_id"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GradeAnswerRequest"
                    }
                },
                "overall_feedback": {
                    "type": "string"
                },
                "submission_id": {
                    "type": "integer"
                }
            }
        },
        "model.BatchGradeRequest": {
            "type": "object",
            "required": [
                "submissions"
            ],
            "properties": {
                "submissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BatchGradeItem"
                    }
                }
            }
        },
...
        "model.GradeAnswerRequest": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "feedback": {
                    "type": "string"
                },
                "question_id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "model.GradeSubmissionRequest": {
            "type": "object",
            "required": [
                "answers"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GradeAnswerRequest"
                    }
                },
                "overall_feedback": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        },
...
    }
}